+++ agent/packaging/deb/mkdeb.sh
@@ -78,10 +78,11 @@ fpm --verbose \
     --maintainer "${DEBFULLNAME} <${DEBEMAIL}>" \
     --license PostgreSQL \
     --url http://temboard.io/ \
-    --depends python3-pkg-resources \
-    --depends ssl-cert \
-    --depends 'python3-psycopg2 >= 2.7' \
     --depends python3 \
+    --depends python3-cryptography \
+    --depends python3-pkg-resources \
+    --depends 'python3-psycopg2 >= 2.7' \
+    --depends ssl-cert \
     --after-install share/restart-all.sh \
     "$@" \
     ./=/