+++ dev/agent-entrypoint.sh
@@ -8,4 +8,10 @@ pip cache dir
 pip install -e /usr/local/src/temboard/agent/ psycopg2-binary hupper
 chown -R "$(stat -c "%u:%g" "$0")" /usr/local/src/temboard/agent
 
+if ! [ -f /etc/temboard-agent/signing-public.pem ] ; then
+	# Copy dev signing key to prefetch UI key.
+	mkdir -p /etc/temboard-agent/
+	cp -v /usr/local/src/temboard/dev/signing-public.pem /etc/temboard-agent/
+fi
+
 exec /usr/local/src/temboard/agent/docker/entrypoint.sh "$@"