+++ ui/temboardui/cli/app.py
@@ -21,10 +21,11 @@ from ..toolkit.app import (
     BaseApplication,
     define_core_arguments,
 )
-from ..toolkit.configuration import OptionSpec
+from ..toolkit.configuration import OptionSpec, MergedConfiguration
 from ..toolkit.errors import UserError
 from ..toolkit.proctitle import ProcTitleManager
 from ..toolkit.services import Service
+from ..toolkit.signing import load_private_key
 from ..toolkit.tasklist.sqlite3_engine import TaskListSQLite3Engine
 from ..version import __version__, format_version, inspect_versions
 from ..web import Error404Handler, app as webapp
@@ -49,6 +50,10 @@ class TemboardApplication(BaseApplication):
         'maintenance',
         'statements',
     ]
+
+    def __init__(self, *a, **kw):
+        super(TemboardApplication, self).__init__(*a, **kw)
+        self.config = TemboardUIConfiguration()
 
     def define_arguments(self, parser):
         define_core_arguments(parser)
@@ -153,6 +158,27 @@ class TemboardApplication(BaseApplication):
         )
 
 
+class TemboardUIConfiguration(MergedConfiguration):
+    def __init__(self, *a, **kw):
+        MergedConfiguration.__init__(self, *a, **kw)  # PY2, use super.
+        self._signing_key = None
+
+    @property
+    def signing_key(self):
+        # Lazy load signing key.
+        if not self._signing_key:
+            path = self.temboard.signing_private_key
+            logger.debug("Loading signing key from %s.", path)
+            try:
+                fo = open(path, 'rb')
+            except OSError as e:
+                raise UserError("Failed to laod signing key: %s" % e)
+
+            with fo:
+                self._signing_key = load_private_key(fo.read())
+        return self._signing_key
+
+
 def bootstrap_tornado_app(webapp, config):
     webapp.config = config
 
@@ -190,7 +216,12 @@ def finalize_tornado_app(webapp, config):
         }),
         (r"/fonts/(.*)", tornado.web.StaticFileHandler, {
             'path': base_path + '/static/fonts'
-        })
+        }),
+        # Path needs a (unused) path parameter, not used by subclass
+        # SingleFileHandler.
+        (r"/(signing.key)", SingleFileHandler, {
+            'path': config.temboard.signing_public_key,
+        }),
     ]
 
     # Append rules *after* plugins because plugins shares same namespace for
@@ -282,6 +313,18 @@ class TornadoService(Service):
             autoreload.watch(path)
 
 
+class SingleFileHandler(tornado.web.StaticFileHandler):
+    @classmethod
+    def get_absolute_path(cls, root, *a):
+        return root
+
+    def validate_absolute_path(self, root, absolute_path):
+        if not os.path.exists(absolute_path):
+            logger.warning("Inexistant file %s", absolute_path)
+            raise tornado.web.HTTPError(404)
+        return absolute_path
+
+
 class VersionAction(_VersionAction):
     def __call__(self, parser, *_):
         print(format_version().strip())
@@ -316,6 +359,12 @@ def list_options_specs():
         s, 'ssl_key_file',
         default=OptionSpec.REQUIRED, validator=v.file_)
     yield OptionSpec(s, 'ssl_ca_cert_file', validator=v.file_)
+    yield OptionSpec(
+        s, 'signing_private_key',
+        default='signing-private.pem', validator=v.path)
+    yield OptionSpec(
+        s, 'signing_public_key',
+        default='signing-public.pem', validator=v.path)
     yield OptionSpec(s, 'cookie_secret', validator=cookie_secret)
     home = os.environ.get('HOME', '/var/lib/temboard')
     yield OptionSpec(s, 'home', default=home, validator=v.writeabledir)